+++ src/components/Pagination/Pagination.module.scss
@@ -1,28 +1,12 @@
 @use "@/styles/globals.scss" as *;
 
-.paginationContainer {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-top: 60px;
-  position: relative;
-
-  @include Size("mobile") {
-    margin-top: 40px;
-  }
-}
-
 .pagination {
   display: flex;
   justify-content: center;
   align-items: center;
-  gap: 8px;
+  gap: 6px;
 
   @include Drag;
-
-  @include Size("mobile") {
-    gap: 4px;
-  }
 
   button {
     width: 36px;
@@ -58,9 +42,4 @@
   .active {
     background-color: $gray20;
   }
-}
-
-.uploadBtn {
-  position: absolute;
-  right: 0;
 }
